fix highlevel output for trb_register_read_mem()
The linear data stream in the response to a trb_register_read_mem()
request has a header field with the length of data for each endpoint's
response. This change fixes the interpretation of those data streams
if endpoints are not responding with the expected amount of data
(for example because they don't offer the requested register).
In the latter case, the structured data (dictionary) returned by the
high level implementation TrbNet.trb_register_read_mem() will
contain a key for the responding endpoint but the corresponding data
will be an empty list.