remove trb_ prefix from high-level TrbNet() methods
The high-level TrbNet class now doesn't shadow the low level methods any
longer. An example: the C function trb_register_read() from libtrbnet.so
is availabe in the low-level Python class asi _TrbNet.trb_register_read().
The high-level class TrbNet (without the underscore) implements a method
TrbNet.read_register() (new: without the trb_prefix)  which returns
structured Python output (a dict):

    >>> import trbnet
    >>> t = trbnet.TrbNet()
    >>> t.register_read(0xfe4d, 0x0)
    {53264: 473956352, 53265: 440401920, 53267: 445644800}
    t.trb_register_read(0xffff, 0x0)

Now, calling the underlying low-level equivalent
_TrbNet.trb_register_read() can be done if needed:

    >>> t.trb_register_read(0xfe4d, 0x0)
    [53264, 475004928, 53265, 439353344, 53267, 445644800]

Before this changeset, the method was only available using the super()
function:

   >>> super(trbnet.TrbNet, t).trb_register_read(0xfe4d, 0x0)
   [53264, 472907776, 53265, 440401920, 53267, 445644800]

BTW: The trb_ prefix is not needed in the end, because the methods are
encapsulated in the namespace of their class. This is already true for
the (low-level) _TrbNet class but there we keep the trb_ prefix in order
to match the names of the underlying C library libtrbnet.so.