Custom TrbException and TrbError based on IntEnum
TrbError is an enum giving meaningful names to integer error
values ( as returned by _TrbNet.trb_errno() ).
Unfortunately, we cannot fetch the enum name strings from the
libtrbnet library. They do not make it into the compilation output.
We now rather reproduce the enum defined in
trbnettools/libtrbnet/trberror.h
It should be synchronized from time to time with the upstream
source code.
So this commit partially reverts d50071e, which removed a
dictionary containing the mapping errno -> enum names.

Based on the TrbError, this commit also introduces a custom
exception called TrbException. It's meant to be raised by
the lowlevel _TrbNet class when any interaction with TrbNet fails.